4032 - Common component Avatar
