--- src/client/pages/CountryHome/Collaborators/UserList/Users/UserCard/Avatar/Avatar.tsx
@@ -1,19 +1,0 @@
-import './Avatar.scss'
-import React from 'react'
-
-import { CountryUserSummary, Users } from 'meta/user'
-
-type Props = {
-  user: CountryUserSummary
-}
-
-const Avatar: React.FC<Props> = (props: Props) => {
-  const { user } = props
-  return (
-    <div className="user-avatar">
-      <img alt={user.fullName} src={Users.profilePictureUri(user.id)} />
-    </div>
-  )
-}
-
-export default Avatar
